Day 1-2.1
This is more general and compresses the else if into a single line. I'd say that's even more elegant, no?

On the other hand, from a performance standpoint it's probably overkill to use whatever sort is used in Array.sort(). This is one of the funny cases where Bubble sort would be faster. I don't really care to optimize this though, so it's written here instead of actually implemented.