Day 1-2.1
Added a leaderboard for day 1 as well, cuz why not.

Turns out Elf 196 is a chad in my input, bringing a whole one (1) snack that is worth 64103 calories. Seriously, did you bring a pure uranium banana or something?

The formatting sucks by the way. Need to find a way to not have to stuff all of the formatting into one line...

Oh, I also changed the name of updateCaption to displayCaption for naming consistency with displayText.